+++ src/app/api/dashboard/tags/[slug]/route.ts
@@ -2,17 +2,25 @@ import { prisma } from '@/lib/prisma'
 import { NextRequest, NextResponse } from 'next/server'
 
 export async function DELETE(req: NextRequest, { params }: { params: Promise<{ slug: string }> }) {
-	const slug = (await params).slug
 	try {
+		const slug = (await params).slug
+		if (!slug) {
+			return NextResponse.json({ message: 'slug parameters is missing' }, { status: 400 })
+		}
+		const existingTag = await prisma.tag.findUnique({
+			where: { slug },
+			select: { id: true },
+		})
+
+		if (!existingTag) {
+			return NextResponse.json({ message: `Tag with slug '${slug}' not found` }, { status: 404 })
+		}
 		const tag = await prisma.tag.delete({
 			where: { slug },
 		})
-		if (!tag) {
-			return new Response('tag not found', { status: 404 })
-		}
-		return NextResponse.json({ message: 'tag deleted successfully' }, { status: 200 })
+		return NextResponse.json({ tag }, { status: 200 })
 	} catch (error) {
 		console.error('Error deleting tag:', error)
-		return NextResponse.json({ message: 'Error deleting tag' }, { status: 500 })
+		return NextResponse.json({ error }, { status: 500 })
 	}
 }
